requirements.txt
typer[all]>=0.9.0
jinja2>=3.1.2
pyyaml>=6.0.1
rich>=13.0.0
